requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
python-dotenv==1.0.1
redis==5.2.1
aiohttp==3.11.11
APScheduler==3.10.4
pytz==2024.2
